+++ lib/src/index.ts
@@ -10,3 +10,4 @@ export * from './id3.js';
 export type * from './request.js';
 export * from './structuredfield.js';
 export * from './utils.js';
+export * from './cea.js';